moved ScreenManager and ScreenEnum to own package for more readability
